feat: add preselected packages setting to limit test preselection
This introduces a new client-scoped string setting containing a
comma-separated list of module IDs. If that list is non-empty, only
batches registered with their `preSelected` set to or using their
default value of `true` belonging to a package whose ID is included in
this list are effectively treated as preselected for the purposes of UI
preselection and automatic test running on startup, should that setting
be enabled.

See-also: #117
Closes: #133