feat: enable batch key filtering in settings and `quench.runBatches`
This allows `quench.runBatches` to run not only with specific batch
keys, but also with strings containing wildcards, so that e.g. all
batches starting with `quench.` could be run with `quench.**`.

This renders `runAllBatches` obsolete, and replaces the old setting for
preselected packages for one doing the same job, but with filters.
Old `preselectedPackages` settings are automatically migrated to a new
`preselectFilters` setting.

BREAKING CHANGE: replace preselect packages with preselect filters