feat: add JSON report generation
By default, Quench generates a reports the way Mocha would, and stores
it in `quench.reports.json` as well as firing a `quenchReports` hook
that receives an object containing this reports as `json` as argument.
When calling `quench.runBatches` with options including `json: true` or
additional JSON config options, Quench also uploads the JSON report,
either in the default location `Data/quench-report.json`, or in a user
configured place.